feat: move pageTitle and directoryList component from Layout component to Post template
